--- ocr/final/IMG_0250.jpeg.txt
@@ -1,84 +1,0 @@
-mereo,
-merere, meruL, meritus. to deserve, to earn
-mergö, mergere, mersi, mersus. to submerge, to flood, to
-inundate
-mergus, -i. sea gull, sea bird
-meritus, -a, -um. deserved, merited, earned
-méta, -ae. turning-point; turning post [the post around
-which chariots turned in Roman chariot races; often
-used metaphorically to refer to a turning point or shift
-in events]
-metuö, metuere, metui, metutus. to fear, to be afraid
-metus, -üs. fear
-meus, -a, -um. my, mine
-micö, micäre, micui, to shine, to flash, to gleam
-mihi. [see ego]
-miles, militis soldier
-militia, -ae. military service
-militö, -äre, -ävi, -äturus. to be a soldier, to serve as a
-soldier
-mille (indecl.). one thousand
-Minerva, -ae. [name: the Roman goddess of wisdom, akin
-to the Greek goddess Athena]
-minimus, -a, -um. least, smallest, slightest
-ministerium, -i. activity, duty
-ministra, -ae. maid, female servant
-minium, -i. cinnabar [a vibrant reddish-orange mineral
-that was used as both a dye and a medicine]
-minor, minus (adj.; gen. = minoris). less
-Minös, Minöis [name: legendary king of Cretel
-minuö, minuere, minui, minütus. to decrease, to lessen,
-to diminish
-miräbilis, -e. wondrous, amazing, shocking
-miror, miräri, mirätus. to wonder, to be amazed
-mirus, -a, -um. wondrous, amazing, wonderful
-misceö, miscére, miscui, mixtus. to mix, to blend, to
-mingle
-miser, misera, miserum. miserable, wretched, sad,
-pitiable, awful
-miseräbilis, -e. miserable, distraught
-miserandus, -a, -um. to be pitied; miserable, wretched
-mittö, mittere, misi, missus. to send
-moderätus, -a, -um. moderated, slow, restrained, held
--ul, -Itus. to warn
-monile, monilis necklace
-monimentum, -i. monument, memorial
-mons, montis mountain
-mora, -ae. delay, wait, pause; [sometimes used with a
-negative to mean:] without delay
-morior, mori, mortuus. to die
-moror, moräri, morätus. to delay, to wait, to remain
-mors, mortis death
-morsus, -üs. jaws, mouth, bite, morsel, nibble
-mortälis, -e. mortal, human
-mörum, -i. mulberry, the fruit of the mulberry tree
-mörus, -i (f.). the mulberry tree
-mös, möris (m.). custom, character
-moveö, movere, mövi, mötus. to move
-mox (adv.). soon
-mucrö, mucrönis (m.). the sharp end ofa sword, the point,
-the tip
-multifidus, -a, -um. split, split into pieces, splintered
-multus, -as -um. many
-münus, muneris gift
-murmur, murmuris (n.). murmur, whisper
-mürus» -i. wall
-Müsa, -ae. Muse [ninedivine sisters who oversaw the arts]
-mütö, -äre, -ävi, -ätus. to change, to shift
-mütuus, -a, -um. mutual, common
-myrtus, -i (E). myrtle trees myrtle leaf
-nam. for
-Nape, Napes (f.). [name: the servant of Ovid's girlfriend
-Corinna]
-narrö, -äre, -ävi, -ätus. to tell, to narrate, to say
-nascor, nasci, nätus. to be born
-Näsö, Nasönis (m.). [name: Ovid's cognomen]
-näta, -ae. daughter
-nätälis, -e. of or related to birth, birth
-nätüra, -ae. nature, natural order, essence
-nätus; -i. son
--ne. [indicates a direct yes / no question when attached
-to the first word Of the question; or, when used i" an
-indirect question:] Whether, or
-ne (Conj,). lest, so that...not [the introductory word
-to a negative purpose or noun clause, or indirect
